Added a dependecy to use the Gson object.
Also I have added the logic to convert the Location object to a JSON
String in order to store it in Shared Preferences.  When the application
loads it will automatically look for the location last saved and convert
the JSON string back to the Location Object for use within the
application.